# Optimize slow queries used by Geckoboard widgets.