--- a/app/services/claim_reporter.rb
+++ b/app/services/claim_reporter.rb
@@ -2,64 +2,59 @@ class ClaimReporter
   include ActionView::Helpers::DateHelper
 
   def authorised_in_full
-    claims = Claim::BaseClaim.non_draft
-    authorised_claims_this_month = claims.where{ authorised_at >= Time.now.beginning_of_month }.authorised
+    claims = non_draft_claims.count
+    authorised_claims_this_month = non_draft_claims.where{ authorised_at >= Time.now.beginning_of_month }.authorised.count
 
     {
-      count: authorised_claims_this_month.count,
+      count: authorised_claims_this_month,
       percentage: claims_percentage(authorised_claims_this_month, claims)
     }
   end
 
   def authorised_in_part
-    claims = Claim::BaseClaim.non_draft
-    part_authorised_claims_this_month = claims.where{ authorised_at >= Time.now.beginning_of_month }.part_authorised
+    claims = non_draft_claims.count
+    part_authorised_claims_this_month = non_draft_claims.where{ authorised_at >= Time.now.beginning_of_month }.part_authorised.count
 
     {
-      count: part_authorised_claims_this_month.count,
+      count: part_authorised_claims_this_month,
       percentage: claims_percentage(part_authorised_claims_this_month, claims)
     }
   end
 
   def rejected
-    claims = Claim::BaseClaim.non_draft
-    transitions = ClaimStateTransition.where{ (to == 'rejected') & (created_at >= Time.now.beginning_of_month) }
-    rejected_claims_this_month = transitions.map(&:claim).uniq
+    claims = non_draft_claims.count
+    rejected_claims_this_month = ClaimStateTransition.where{ (to == 'rejected') & (created_at >= Time.now.beginning_of_month) }.count('DISTINCT claim_id')
 
     {
-      count: rejected_claims_this_month.count,
+      count: rejected_claims_this_month,
       percentage: claims_percentage(rejected_claims_this_month, claims)
     }
   end
 
-  def rejected_count
-    rejected_claims = Claim::BaseClaim.where(state: 'rejected')
-    rejected_claims.count
-  end
-
-  def outstanding
-    Claim::BaseClaim.where(state: %w( allocated submitted redetermination )).order(original_submission_date: :asc)
-  end
-
-  def oldest_outstanding
-    outstanding.first
-  end
-
   def completion_rate
-    intentions = ClaimIntention.where{ created_at >= 16.weeks.ago }
-    claims = Claim::BaseClaim.non_draft.where{ created_at >= 16.weeks.ago }
+    intentions_form_id = ClaimIntention.where{ created_at >= 16.weeks.ago }.pluck(:form_id)
+    claims_form_id = non_draft_claims.where{ created_at >= 16.weeks.ago }.pluck(:form_id)
 
-    completed = claims.map(&:form_id) & intentions.map(&:form_id)
+    intentions = intentions_form_id.size
+    completed = (claims_form_id & intentions_form_id).size
 
     claims_percentage(completed, intentions)
   end
 
   def processing_times
-    processing_times = processed_claims.inject([]) do |times, claim|
-      processed_timestamp = claim.claim_state_transitions.first.created_at
-      submitted_timestamp = claim.original_submission_date
-      times << (submitted_timestamp - processed_timestamp)
-    end.sort
+    claims_map = processed_claims.where{ created_at >= 16.weeks.ago }.uniq.pluck(:id, :original_submission_date).to_h
+    claims_states = ClaimStateTransition.where(claim_id: claims_map.keys).order(created_at: :desc).pluck(:claim_id, :created_at)
+
+    timings = claims_states.map do |(claim_id, created_at)|
+      next unless claims_map.key?(claim_id)
+
+      processed_timestamp = created_at
+      submitted_timestamp = claims_map.delete(claim_id)
+
+      (processed_timestamp - submitted_timestamp)
+    end
+
+    timings.compact
   end
 
   def average_processing_time
@@ -74,16 +69,19 @@ class ClaimReporter
   private
 
   def calculate_average(collection)
-    collection.inject{ |sum, e| sum + e }.to_f / collection.size
+    collection.sum.to_f / collection.size
   end
 
   def processed_claims
     Claim::BaseClaim.caseworker_dashboard_completed
   end
 
-  def claims_percentage(percentage_claims, all_claims)
-    return 0.0 if percentage_claims.none? && all_claims.none?
+  def non_draft_claims
+    Claim::BaseClaim.non_draft
+  end
 
-    (percentage_claims.count.to_f / all_claims.count.to_f) * 100
+  def claims_percentage(subset_count, all_count)
+    return 0.0 if subset_count.zero? && all_count.zero?
+    (subset_count.to_f / all_count.to_f) * 100
   end
 end