A much needed optimisation has been done. Here are the results using the same database (8122 claims):

**PRE-REFACTOR**

/geckoboard_api/widgets/claims
* Completed 200 OK in 625ms

/geckoboard_api/widgets/claim_completion
* Completed 200 OK in 25602ms

/geckoboard_api/widgets/average_processing_time
* Completed 200 OK in 24384ms


**POST-REFACTOR**

/geckoboard_api/widgets/claims
* Completed 200 OK in 119ms

/geckoboard_api/widgets/claim_completion
* Completed 200 OK in 70ms

/geckoboard_api/widgets/average_processing_time
* Completed 200 OK in 350ms